Move ScreenShot monitoring code to ScreenShotHelper
Clean up the MainActivity by moving the ScreenShot handling code to a
separate class.
